Update : 테스트 코드 리팩토링 테스트 코드에서 app.play() 메소드만을 실행하도록 수정합니다. generateRandomNumber를 모듈로 분리합니다.
